test: add board test test: add move test refactor: remove some code
