+++ src/__test__/geometry/move.test.ts
@@ -3,7 +3,7 @@ import { calculateMoveListForPawn } from '../../geometry/move.ts';
 import { Fen } from '../../geometry/types.ts';
 import { createChessBoardFromFen } from '../../geometry/board.ts';
 describe('Move generation for pawn', () => {
-  it.skip('generates all the moves for an e2 pawn in init position', () => {
+  it('generates all the moves for an e2 pawn in init position', () => {
     const initPos: Fen = INIT_POSITION;
     const board = createChessBoardFromFen(initPos);
     const moves = calculateMoveListForPawn({ row: 2, column: 'e' }, board);
